Fix adding audio to video bug
After having switched  to the ubuntu based ffmpeg container there was a bug when adding the correlated audio track to the video file. The copied video files had a size of 0kb and an error was thrown.
This is now fixed by using expicitely the appropriate audio codec in bin/03_merge_audio_to_video.sh